// JeeFlatForm-MicroFrontend-/projects/jeehr/src/app/page/JeeHRModule/so-do-to-chuc/Model/so-do-to-chuc.model.ts
export class LyLichNhanVienModel {
	ID_NV: number;
	Image:string;
	MaNV: string;
	HoTen: string;
	GioiTinh: string;
	NgaySinh: string;
	NoiSinh: string;
	NguyenQuan: string;
	DiaChiThuongTru: number;
	DiaChiTamTru: string;
	SoDienThoaiNha: string;
	DiDong: string;
	EmailCongTy: string;
	SosoBHXH: number;
	TinhTrangHonNhan: string;
	MaSoThue: string;
	CMND: string;
	NgayCapCMND: string;
	NoiCapCMND: number;
	TrinhDoHocVan: string;
	TruongTotNghiep: string;
	ChuyenMon: string;
	TrinhDotinHoc: string;
	DanToc: string;
	TonGiao: string;
	TrinhDoThongTin: any[] = [];
	BangCapKhac: any[] = [];
	NguoiThan: any[] = [];
	Structure: string;
	ChucDanh: string;
	NgayBatDauLamViec: string;
	NgayVaoChinhThuc: string;
	QuanLyHienTai: string;
	HopDong: any[] = [];
	TruocKhiVaoCty: any[] = [];
	QuaTrinhCongTac: any[] = [];
	QuaTrinhDanhGia: any[] = [];
	QuaTrinhNghiDaiHan: any[] = [];
	KhenThuong: any[] = [];
	KyLuat: any[] = [];
	clear() {

	}
}